revert(docs): 404エラー解消のため仕様書リンクを元に戻す
- 404エラーが発生していたため、ゲーム画面からの仕様書へのリンク機能を、本ブランチ作成前の状態に戻しました。
- `app/games/[slug]/docs/[doc].tsx` のルーティングは削除し、`app/games/[slug]/rules/page.tsx` を再作成しました。
- `GameLayout.tsx`、`GameClientPage.tsx`、`app/games/[slug]/page.tsx` の関連する変更を元に戻しました。

仕様書の分割 (`spec-action.md`, `spec-display.md`) は維持されています。
