+++ app/games/[slug]/page.tsx
@@ -20,16 +20,7 @@ async function getGameData(slug: string) {
   const manifestPath = path.join(process.cwd(), 'games', slug, 'manifest.json');
   const manifest = JSON.parse(fs.readFileSync(manifestPath, 'utf-8')) as GameManifest;
 
-  const rulesPath = path.join(process.cwd(), manifest.rulesFile.replace(/^\//, ''));
-  const rulesContent = fs.readFileSync(rulesPath, 'utf-8');
-
-  const specActionPath = path.join(process.cwd(), manifest.specActionFile.replace(/^\//, ''));
-  const specActionContent = fs.readFileSync(specActionPath, 'utf-8');
-
-  const specDisplayPath = path.join(process.cwd(), manifest.specDisplayFile.replace(/^\//, ''));
-  const specDisplayContent = fs.readFileSync(specDisplayPath, 'utf-8');
-
-  return { manifest, rulesContent, specActionContent, specDisplayContent };
+  return { manifest };
 }
 
 export default async function GamePage({ params }: PageProps) {
@@ -37,5 +28,5 @@ export default async function GamePage({ params }: PageProps) {
   const { slug } = resolvedParams;
   const { manifest } = await getGameData(slug);
 
-  return <GameClientPage manifest={manifest} slug={slug} rulesFile={manifest.rulesFile} specActionFile={manifest.specActionFile} specDisplayFile={manifest.specDisplayFile} />;
+  return <GameClientPage manifest={manifest} slug={slug} />;
 }
